azure-data-factory/azure-data-factory/function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
ubiops==3.3.0